Adapt example according to latest RunInference changes
